requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.9
pydantic-settings==2.11.0
email-validator==2.3.0
requests==2.32.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.4.2
httpx==0.28.1
python-multipart==0.0.20
